+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 *   (bug) Add empty extra field for slices.
 *   (improvement) Added `ImageValue` and `VideoValue`.
 *   (improvement) Implement `EmbedField` (it only supports videos for now).
+*   (bug) Fix invalid initialization of `SliceValidationCompound`.
 
 
 4.0.0-beta.2