+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 ============
 
 *   (bug) Add empty extra field for slices.
-*   (improvement) Added `ImageValue` and `VideoValue`.
-*   (improvement) Implement `EmbedField` (it only supports videos for now).
+*   (feature) Add `ImageValue` and `VideoValue`.
+*   (feature) Implement `EmbedField` (it only supports videos for now).
 *   (bug) Fix invalid initialization of `SliceValidationCompound`.
+*   (feature) Add `DocumentLinkValue` and use it in `LinkField`.
 
 
 4.0.0-beta.2